requirements.txt
Django==1.10.2
beautifulsoup4==4.5.1
dj-database-url==0.4.1
django-bootstrap3==7.1.0
gunicorn==19.6.0
psycopg2==2.6.2
python-postmark==0.4.10
requests==2.11.1
wsgiref==0.1.2
